requirements.txt
diffusers==0.33.0
transformers==4.40.1
deepspeed==0.15.4
accelerate==0.29.3
onnxruntime-gpu
numpy==1.26.4
opencv-python-headless==4.8.0.74
imageio==2.37.0
imageio-ffmpeg==0.6.0
einops
ninja
pandas
decord
tensorboard
pytorch-lightning
IPython
timm
pandarallel
beautifulsoup4
ftfy
matplotlib
# flash_attn_3 3.0.0b1